Set default initializer for variational distribution's stddev to be close to 0.
The defaults in passing around arguments were incorrect. In particular, originally a `None` initializer is passed to `untransformed_scale` in all mean-field variational distributions; Keras defaults this to Glorot uniform. This is bad for initializing scale parameters: converging to a non-trivial solution typically requires an initialization <-3. This is because its transform implies a stddev initialization close to 0, which allows the parameter update trajectory to begin roughly like non-VI-based gradient descent.

PiperOrigin-RevId: 192850120