Handle deprecated method inspect.getargspec() This PR is related to the issue: [DeprecationWarning: inspect.getargspec() is deprecated #78](https://github.com/tensorflow/probability/issues/78).
As in Python 3.6, inspect.getargspec() was deprecated, and using inspect.signature() or inspect.getfullargspec() is encouraged.

With this PR changes, program_transformations.py uses inspect.getfullargspec() when it's available, i.e. when Python 3 is used, and fallback on inspect.getargspec() otherwise. The attributes in the returned object are slightly different, but we only care about the `args` attribute and it's unchanged.

PiperOrigin-RevId: 203512432