Increment version number 0.2.0 -> 0.3.0
With 0.2.0 released, I believe we should follow the TF convention of
incrementing the version number now. This way, nightly releases will be named
'0.3.0.devYYYYMMDD' which suggests (rightly) that they are dev snapshots in
pursuit of the eventual 0.3 release.

In the future, the release workflow (documentation in progress) will call for
incrementing the version number right after the release branch is cut (or
thereabouts...specifics TBD).

PiperOrigin-RevId: 205444185