Stop building source distributions.
They're not really needed, since our distributed package should work with any
python version on any platform, and it's not clear that their installation will
work correctly in the future, should we, say, include .cc or .so contents in
the distro. Currently, it's just deadweight on pypi and clutter in our release
configs.

PiperOrigin-RevId: 205472513