Support zero batch dims in RealNVP Bijector.
The default implementation of shift_and_log_scale_fn uses tf.layers.dense,
which in turn assumes that inputs are of shape `[b1, ..., bB, F]` where
`B >= 1`. The RealNVP implementation was written so as to encode this
assumption as well, but it need not be. The fix is simply to change the slice
operations to use, e.g., `[..., :N]` ("all of the zero or more dims up to the
last") instead of `[:, :N]` ("all of the *one* or more dims up to the last").

A test is also added, to cover both batched and non-batched inputs, but this,
of course can't use tf.layers.dense as the transformation. We keep the existing
test (renamed to indicate its non-linear xform) and add a new one with a
trivial `x \mapsto (x, x)` transformation.

PiperOrigin-RevId: 207164580