Added default name for TransformedTransitionKernel and various tests improvements.
Removed test about the acceptance_ratio in hmc_test.py that passed only due to luck.
Modified test in hmc_test to ensure that the state does not leave the support of the distribution, via the introduction of a transformed kernel.
Corrected tests in langevin_test.py, where the true_mean parameter was twice accounted for in the Gaussian distributions, and added non trivial values to true_mean so that it does not go unnoticed again.

PiperOrigin-RevId: 208460393