Added example notebooks for TFP
This PR contains notebooks that are inspired by the Edward1 Example
notebooks, but the models have been reimplemented in Tensorflow
Probability. The main purpose of this PR was to create placeholder links
so the notebooks themselves could be linked to GitHub (these will be
changed when the final notebooks are put in a PR). Many of these
notebooks are still in progress.