Override _covariance method in GP and GP regression Distributions.
MVNLinOp, of which the GP distributions are subclasses, defines these as
`scale * scale^T` which puts an extra, unnecessary Cholesky op in the graph
dependencies of the covariance matrix. Some calculations (e.g., for sparse GP
methods) may wish to know these matrices but explicitly avoid the O(N^3)
operation. Thus, in this change we switch to capturing them in `__init__`,
upstream of any cholesky ops, and override the _covariance method to return
these values.

PiperOrigin-RevId: 211121975