Example for CIFAR-10 with ResNet18 and VGG16 This is an example for CIFAR-10 with ResNet18 and VGG16. With the current setup, the ResNet18 hits about 80% validation accuracy and 90% training accuracy.
The visualization for MNIST has been removed because the plots did not really look as I expected. The validation accuracy would be pretty high but the distributions plotted are still like a random dist. I'm unsure as to why that is happening so I removed it for now. If you would like it for debugging, I can include it.

Please let me know if there are style changes needed.

EDIT (@dustinvtran): I completed the minimum to pass linting errors. There are a few other style changes worth doing in a separate change.

PiperOrigin-RevId: 213055332