Fix off-by-one error for linear gaussian SSMs with time-varying transitions.
We define `transition_matrix(t)` as the matrix implementing the linear transition from timestep `t` to timestep `t+1`. This was implemented correctly in `forward_filter`, but the mean/sample/variance methods were implicitly treating it as the transition from `t-1` to `t`, which is incorrect. This CL fixes the error, clarifies the semantics in the class docstring, and adds tests.

PiperOrigin-RevId: 213568851