Make RealNVP tests less tolerant of floating point discrepancy.
The RealNVP tests were originally written with relative tolerance of .1; this
is quite high. The reason this tolerance was needed is that the test inputs
were chosen to be mostly well outside the well-behaved region of the NVP
bijector. The test inputs were taken from, eg, np.arange(N) for N an integer
of magnitude 10^1. Meanwhile, typical inputs to a RealNVP bijector are drawn
from Normal(0, 1), so these test inputs are quite unrealistic for standard
operation.

This change modifies the test inputs to be drawn from np.random.normal(0, 1),
and updates the tolerances to be tighter. We still get failures at rtol=1e-5,
but at 1e-4 no flakes were observed across 100 test run repetitions.

PiperOrigin-RevId: 214312270