Make tfb.Reshape work correctly with tfb.Chain.
Previously, this combination would not compute the event_ndims correctly in cases where Reshape changed the rank of the event (and possibly other cases).

This required setting the forward/inverse_min_event_ndims. Unfortunately,
arriving at those values requires that the rank of the input shapes must be
statically known, due to the way tfb.Bijector uses those values. In the future
we might be able to relax this requirement.

PiperOrigin-RevId: 223220501