Allow MaskedAutoregressiveFlow bijector to work with tensor-valued events.
While this could have been emulated by careful reshaping back and forth (as is
done in the added unittests), this was highly unergonomic and not actually
required by the algorithm.

Additionally, clarify in the documentation that this bijector does not actually
enforce a particular order of autoregressivity, and the user is free to use
something akin to orderless-MADE to implement the shift+scale callback.

PiperOrigin-RevId: 224164341