Eliminate _prob, _log_prob, _cdf, _log_cdf stubs in Distribution.
The origin of this change was github.com/tensorflow/probability/issues/233. In order to fix it, we needed to ensure the proper routing of prob/log_prob calls when Distributions wrapped in meta-Distributions only implemented one or the other. Ultimately this meant that meta-Distributions needed to know whether wrapped Distributions had implemented one or the other. Having stubs in the base Distribution class prevented detecting this.

Eliminating the stubs exposed issues in how the Independent meta-Distribution was deferring its prob/log_prob/cdf/log_cdf calls to the wrapped Distribution; we've therefore adapted it to this change, as well.

Fixes: #233
PiperOrigin-RevId: 224219903