Return static `event_shape.ndims` from Independent distributions even when `distribution.batch_shape.ndims` is not known.
Note: in some cases this means we'll return a static value for `event_shape.ndims` even when we'll later discover that the dynamic batch shape is invalid, i.e. has `batch_shape.ndims < reinterpreted_batch_ndims`. IMHO this is fine since the static ndims we return is the only value that could be valid (i.e., conditioned on your program not crashing in the future, this is the correct value).
PiperOrigin-RevId: 225583144