Make Invert Bijector call public methods instead of private ones.
I am increasingly certain this should be the rule in general, for any meta-Foo
(meta-Distribution, meta-Bijector, meta-PositiveSemidefiniteKernel). The need
for a pylint override should always be a hint that we're doing something fishy.
In this case, I needed to make this change to correctly address a caching issue
in the base Bijector class. There was some whacky ping-ponging going on down
there, where the actual functioning of Invert depended on the base Bijector's
try/except logic for figuring out which methods to use. I'm in the process of
eliminating that try/except logic in favor of explicit existence checks and
this class's overriding of the private methods was causing problems for that
change.

PiperOrigin-RevId: 226197773