Move monte_carlo and trainable_distributions into submodules.
(This is a non-user-facing change; we just shuffled some files around).

Having these as top-level files is a) inconsistent with everything else and b)
causes zany import issues in certain circumstances. Specifically, Travis CI
uses a stripped-down python binary that doesn't include `math` as a built-in
module, but instead loads it from a .so file somewhere. This means that any
module called `math` that is in the same directory as a script (e.g., our
monte_carlo_test.py) gets imported by anything in that script's import tree
that does an `import math`. I.e., this causes our math package to be imported
instead of stdlib math, which is...not good.

Long story short, moving these modules makes things more consistent and fixes
the issue.

PiperOrigin-RevId: 226233986