Bugfixes and tests for non-float32 dtype and dynamic shape in STS fitting.
This fixes an oversight in the original test cases. Adding tests for float64 and dynamic shape turned up a couple of issues, which this CL fixes:
 - get_variable() defaults to dtype=float32 in some circumstances if using lambda initialization and no static dtype is specified, so we need to specify a static dtype. (`param.prior.dtype` is not *quite* right -- for an int-valued param parameterized by float32 logits it'd return tf.int32 instead of tf.float32 -- but we don't support int-valued params anyway, so I think it's 'good enough' for current purposes)
 - initializing a variable based on the shape of another variable fails when the first variable's initialization is dynamic. So we have to be a little bit inefficient and initialize `loc` and `scale` separately.

In debugging this I started thinking about a broader refactor of some of this code to fit with a more object-oriented, tf2.0-style approach. I'm making some progress on that in cl/225630712, which will hopefully subsume these fixes eventually.   This CL is meant as a 'minimum viable' fix for the current API.

PiperOrigin-RevId: 226382479