Fix Bijector caching of constant log-det-jacobian.
Caching was causing spurious graph dependencies between otherwise independent
ops, caused by caching of LDJ when the first value was passed through. See
added unit test for details.

The fix is a rather involved refactoring of the base Bijector class logic. I *think* it's rather cleaner at the end of the day, and a bunch of code duplication was eliminated.

Fixes: #253
PiperOrigin-RevId: 226383030