Allow the variational flipout layers to work with 'channels_first' data format.
Fixes https://github.com/tensorflow/probability/issues/154.

The fix itself is straightforward, the main issue is testing this change. It
turns out that on the CPU, TF's convolutions don't support the 'channels_first'
data format. As a workaround, I monkey-patched a convolution implementation
that standardizes the format to 'channels_last', but still lets us test the
relevant functionality of our layers.

Along the way I removed the unnecessary tiling from the flipout layers.

PiperOrigin-RevId: 227903491