Use scipy.special functions when testing KL divergence for Chi, Chi2.
This requires some small decrease in relative tolerance for true_kl vs. kl_. This is justified because whereas before we were comparing a tensorflow computation to a tensorflow computation, we are now comparing a tensorflow computation to another library (scipy) so there might be some expected delta based on e.g. a different computational method.

PiperOrigin-RevId: 227918553