Fix broadcasting bug in the half-Cauchy distribution's log_prob and log_cdf functions.
tf.where(condition, x, y) expects x and y to have the same shape if they are both non-None. The caller must ensure this is the case by broadcasting if required. This CL adds a missing broadcast in the _extend_support_with_default_value function and unit tests to cover the broadcasting case.

PiperOrigin-RevId: 228205852