Increase uncertainty of default STS priors on level/effect parameters.
The goal of this change is to make the default priors sufficiently noninformative that zero is always a plausible output. This is necessary in case the observed time series is partly or entirely explained away by other model components.  The previous priors denied this possibility, leading to pathological inferences when the observed time series has large values and the models contains multiple components.

PiperOrigin-RevId: 231620290