Expose HMC parameters via kernel results.
This enables wrapper kernels to adapt HMC's parameters on the fly. See go/mcmc-adaptation-via-composition for the design.

This is exposed via a constructor argument which is False by default, for
backwards compatibility. The logic here is that the (newly written) wrapper kernels will force this parameter to True thus making the new code look reasonable, while keeping old code functioning as well.

Nothing is deprecated yet due to a lack of replacement
(which is coming in the future changes).

PiperOrigin-RevId: 234192909