Fix and re-enable hmc_test.py
The primary issue was that the step size in the MCEM test was set too high,
causing resonance in the HMC dynamics. This caused extreme sensitivity to
the details of the optimization algorithm.

Once that was fixed, this uncovered a difference in the seed dependence of the
data generating process between TF1 and TF2, so I rewrote it in numpy (it was
shorter anyway).

Also cleaned-up the likelihood function, it was a little too dense.

PiperOrigin-RevId: 234703101