Don't apply a tf.where on unmodified kernel results in the MH kernel.
As a surface level benefit, this reduces the graph size/amount of computation.

The actual reason for this change is that it lets transition kernels have
fields in the kernel results which are not batched as long as they are
read-only. A concrete example are the propagated HMC parameters, like
step_size.

It's not obvious what to do with such parameters if they do change (since
accept/reject is a per-batch index decision, but when they are constant, the
solution is obvious.

There is a small caveat that the method used to detect constness breaks if the
kernel creates a new tensor (e.g. using tf.identity or tf.convert_to_tensor). I don't think this is a big issue, but something an TransitionKernel implementor will need to keep in mind. This also interacts interestingly with tf.function annotations on the one_step method, but I have a plan for that as well.

PiperOrigin-RevId: 235198213