Refactor reshape bijector so we can eventually `tf.function` decorate it.
Because of b/124240153 we must use the `map(tf.identity, deps)` workaround for control deps. However, since we cant tf.identity wrap assertions this means we needed to wire up the control deps against the assertion embeded state, `event_shape_in` and `event_shape_out`.  This in turn required refactoring which spun out of control because of the inherent complexity of this bijector.

PiperOrigin-RevId: 235355363