Update sts.fitting test for TF2 compatibility.
The issue (diagnosed by cgs@, thanks Chris!) was that the `tf.while_loop` in _minimize_in_graph returned inconsistent types: a Tensor in graph mode and a length-one list in Eager mode. When the result was passed into `control_dependencies`, in TF2 it sees a list-of-lists and chokes. (not sure why we didn't see this problem in TF1 eager, but this fix makes the point moot).

The fix is to use `return_same_structure` to ensure that the while loop returns a length-one list (matching the input `loop_vars`) in both graph and eager mode, from which we extract the Tensor op to return.

PiperOrigin-RevId: 235829705