`tf.function` decorate `tfp.bijectors`. Notable Features: - This change preserves caching even in eager mode (because the EagerTensor is cached, not the graph lifted Tensor). - Users can disable `tf.function` by setting `b._use_tf_function=False`. (We can expose this as a "public" setter if requested.)
Alternative Considered:

```python
    _INTERCEPTED_METHODS = {
        '_inverse',
        '_forward',
        '_inverse_log_det_jacobian",
        '_forward_log_det_jacobian",
    }

    def __getattribute__(self, name):
      if name not in _INTERCEPTED_METHODS:
        return object.__getattribute__(self, name)
      if not hasattr(self, name):
        raise NotImplementedError("{} not implemented".format(name))
      attr = object.__getattribute__(self, name)
      if self._use_tf_function:
        return tf.function(attr, autograph=False)
      return attr
```

Q:Why not use this technique?

A1: Overriding `__getattribute__` can be very confusing to the uninitiated.

A2: For understandability, I found I still wanted to add the comment "Maybe decorated." when using this technique. (Next to all of the calls to `{'_inverse', '_forward', '_inverse_log_det_jacobian', '_forward_log_det_jacobian'}`.) This desire indicated (to me) that an explicit call is preferable.
A possible reason to prefer the `__getattribute__` call is that all calls to _INTERCEPTED_METHODS can be decorated, viz., those calls which we dont control from `Bijector` base.  However I do not believe this buys anything because the only calls that "matter" are those which are part of the public API. For any other non-API functions a subclass adds, a user is in general "on their own" anyway.

Alternative Considered:

```python
  def _maybe_tf_function(self, attr_name):
    """Possibly wraps `callable` `f` with `tf.function`.

    Note: This function exists (rather than inlining) so a subclass can choose
    different behavior, eg, `autograph=True`.

    Args:
      attr_name: Python `str` which corresponds to `self.attr_name` which will
        be wrapped by `tf.function` when `use_tf_function` is `True`.

    Returns:
      wrapped_fn: Python `callable` with same semantics as input `fn` but
        possibly wrapped by `tf.function`.
    """
    if not self._use_tf_function:
      return getattr(self, attr_name)
    fn = self._tf_fns.get(attr_name, None)
    if fn is None:
      fn = tf.function(getattr(self, attr_name))
      self._tf_fns[attr_name] = fn
    return fn
```

Q:Why not use this technique?
A:Employs instance level caching of the tf.function wrapped function.
PiperOrigin-RevId: 236670180