Add `constant_offset` option to STS Sum models, and set it to the observed mean by default.
This implements the general idea of centering the observations before fitting them (`observed_time_series = f(x) + constant_offset` is equivalent to `observed_time_series - constant_offset = f(x)`), which is generally sensible and especially important for numerical stability when observations are large. By encoding the mean in the generative process, we guarantee that forecasts and similar derived quantities can't 'forget' to add it back in.

PiperOrigin-RevId: 237497842