Fix a bug introduced when creating forecasts with a constant offset and batch shape.
The bug occurs because we (currently, as part of an unfortunate hack) permute the dimensions of the model parameters during forecasting. It turns out we also need to modify the dimensions of the constant offset when forecasting with a `tfp.sts.Sum` model.
See the comment in `forecast.py` for details.

This change also updates `forecast_test` to include the offending case.

PiperOrigin-RevId: 238270068