Update AdditiveSSM to ignore batch shapes from unused component initial_state_priors.
This fixes a shape error in forecasting, when the components initial state priors have nontrivial batch shapes. This case happened not to come up in `forecast_test`; this CL also fixes that omission.

The specific error was that AdditiveStateSpaceModel was constructing its observation matrix using the full broadcast batch shape of the component models, rather than the (more precise) broadcast batch shape of their observation matrices.

PiperOrigin-RevId: 238275810