Make TransformedTransitionKernel copying work.
Previously the outer kernel was modifying the inner kernel's target log prob in place, meaning that upon every copy it'd get wrapped with an additional layer of the transformation logic. Now the kernel passed into the outer kernel is left unmodified, so that wrapping will apply only to that original kernel.

PiperOrigin-RevId: 239429899