Don't attempt to add chain batch shape padding to time series that don't need it.
When running inference with multiple chains, time series with sample shape need to be padded to account for the new batch dimensions. But time series *without* sample shape don't need any padding (broadcasting will 'just work'), and attempting to do the padding can cause a shape error, so we shouldn't do it.

This CL also moves `pad_batch_dimension_for_multiple_chains` to the internal utils file. This is just to let us unit-test it without exposing it publicly.

PiperOrigin-RevId: 240304317