Fix bug in GPRegressionModel; we were not correctly handling empty data.
The GPRM distribution handles a special case in which the observation data is
either not given (None) or has 0 actual examples in it. We were mishandling the
latter case, and the test was wrong to match. Specifically, the code was
looking at the left-most *feature* dim instead of the "number of examples" dim
(just left of that), and, correspondingly, the test had a zero in the feature
shape, instead of in the number of examples dim. This change fixes both errors.

PiperOrigin-RevId: 241535868