Introduce SchurComplement PSD Kernel
This is a meta-kernel whose `matrix` function acts like the Schur complement.

Given a block matrix  `M = [[A, B], [[C, D]]`, we define the Schur complement of D in M as
`M / D = A - B @ D^-1 @ C`.

Now, suppose we have a kernel `k` and some fixed collection of inputs
`[z0, z1, ..., zN]`. Given new inputs `x` and `y`, we can form a block matrix

  ```none
    M = [
      [k(x, y), k(x, z0), ..., k(x, zN)],
      [k(z0, y), k(z0, z0), ..., k(z0, zN)],
      ...,
      [k(zN, y), k(z0, zN), ..., k(zN, zN)],
    ]
  ```

We might write this, so as to emphasize the block-ness,

  ```none
    M = [
      [xy, xZ],
      [yZ^T, ZZ],
    ],

    xy = k(x, y)
    xZ = [k(x, z0), ..., k(x, zN)]
    yZ = [k(y, z0), ..., k(y, zN)]
    ZZ = "matrix of k(zi, zj)'s"
  ```

Then we have the definition of this kernel's apply method:

`schur_comp.apply(x, y) = xy - xZ @ ZZ^{-1} @ yZ^T`

and similarly, if x and y are collections of inputs, the corresponding matrix
output.

Notably, this kernel creates a Cholesky bijector
(Invert(CholeskyOuterProduct)), for solving ZZ systems, which means we get
caching of cholesky's for free after the first one.

RELNOTES: Introduce SchurComplement PSD Kernel
PiperOrigin-RevId: 242520588