Use functional style for `tensor.dtype.*` thus making TFP code slightly more agnostic to NP or TF.
Main workhorse:
sed -i -e 's/\([ =(]\)\([_a-zA-Z0-9.]\+\)\.dtype\.\([_a-zA-Z0-9]\+\)/\1dtype_util.\3(\2.dtype)/g' *.py

Note: this cl makes a number of code pattern unifications. Because of this it
would be easy to undo via sed.

Note: after this cl we could consider going one step further and having these functions take any object which, if it has a `dtype` member returns the function evaluated on this member or otherwise presumes the input is a dtype. (Meaning such a change is backwards compatible.) This would further simplify the call-side and eliminate the double use of the word `dtype`. This change could be easily done via:

sed -e 's/dtype_util\.\([a-z_]\+\)(\([.a-zA-Z0-9]\+\).dtype)/dtype_util.\1(\2)/g' *.py
# And one variant to handle `base_equal`.

In that world, we may or may not also want a `dtype_util.dtype(x)` for consistency. In so doing we'd never see `x.dtype` accesses in TFP and would be able to easily grep for such deviations, perhaps as a presubmit. (I do not think a presubmit is warranted however as we'd anyways have a TAP run for some other backend were it to exist.)

PiperOrigin-RevId: 243128285