Allow PSDKernels to have dtype "None".
(unrelated style changes: use 'tfpk' consistently as shorthand for tfp.positive_semidefinite_kernels throughout the test code).

Prior to this change, psd kernels fell back to a default dtype of float32.
This was the case, even when they had no need to even specify a particular
dtype. For example, the ExpQuad kernel with no args, when applied to inputs of
any dtype, is able to produce a valid output of the same dtype -- it ignores
its parameters when they are all `None`. However, it still "thinks" it has
dtype float32, which causes issues with various meta-kernels. This came up in
practice with SchurComplement, which tries to infer its dtype from a
combination of the `base_kernel` and the `fixed_inputs` args. The following led
to a type error:

SchurKernel(ExponentiatedQuadratic(), fixed_inputs=np.ones([3, 1], np.float64))

Because the ExpQuad kernel was defaulting to type float32 (for no good/useful
reason), while the explicit type on `fixed_inputs` was float64.

This change makes it so that kernels can have dtype=None when there's no reason
to actually declare one. New tests are added, which fail without the changes to
the kernel impls and pass after the changes.

PiperOrigin-RevId: 243289392