Add internal utilities for dealing with nested args and functions.
These should be used whenever taking user-provided callables and their args.

The overall goal of this CL is to extend the current calling convention (see
the doc) to handle arbitrary nested structures in an ergonomic way. Roughly
speaking, the rule is that given `fn` and `args`, we want to do:

`fn(*args)` if `args` is a sequence.
`fn(**args)` if `args` is a mapping.
`fn(args)` otherwise.

Unfortunately, this logic (arguably) mishandles `namedtuples`, which are
sequences. Nobody wants them to be expanded, so we had to add special logic to
detect them, and pass them in unmolested.

As a companion change, this adds a way to convert `args` to a potentially
nested set of `Tensor`s. This was done slightly differently than one might
expect in that `Tensor` conversion happens eagerly (i.e. the first valid
sub-tree that can get converted to a `Tensor`, is). This too ran afoul of the
`namedtuple` issue (`tf.convert_to_tensor` happily convers `namedtuple`s to
naked `Tensor`s). Unfortunately, to work-around this required a blacklist which
forced our hand with the `call_fn` implementation.

All expansion/recursion has an undocumented escape hatch. If you add a
`_tfp_nest_expansion_force_leaf` class attribute to your type, it'll get passed to `fn` as
`fn(args)` and passed to `tf.convert_to_tensor` as is (useful if you have a
custom tensor conversion function).

PiperOrigin-RevId: 244291526