Fixed evaluation order of distributions within `JointDistributionNamed` triggered when one random variable was used as input to multiple distributions. This includes: 1. modifying topological sort within `_best_order` 2. simplifying method for providing arguments to distribution functions
In particular, the parent class `JointDistributionSequential` uses
an array of samples that looks a lot like a stack frame and uses the
position of an argument to indicate how far down the "stack" you need to
reach to get the arguments supplied to the distribution functions. This
machinery is appropriate to the positional style of the interface to JDS.

`JointDistributionNamed` reuses the machinery. But everything is
referred to by name. So the original version of `_best_order` doesn't
just topologically sort, it also mimics register allocation in a compiler
by translating names into positions on the stack and tries to do so in
a way that minimises how much stack frame each function has to see. But
it doesn't need to do this because the distribution functions provided
to JDN refer to arguments by name. No padding is needed and everything
can be simplified just by using kwargs.

PiperOrigin-RevId: 247289671