Introduce FeatureTransformed psd kernel and some subclasses.
Many methods using psd kernel entail some kind of parameterized transformation
to the inputs. Examples include "automatic relevance determination" (ARD) in
the Gaussian process literature, and the input warping method for Bayesian
optimization in Snoek, et al, 2014.

This CL introduces a meta-kernel, `FeatureTransformed`, which wraps a simple
kernel and applies a user-provided transformation to the inputs before
deferring to the wrapped kernel. This CL also includes 2 particular
transformed kernels, `FeatureScaled` and `KumaraswamyTransformed`, which
correspond to the aforementioned ARD and "input warping" use cases.

PiperOrigin-RevId: 249269800