Define forward log det jacobian explicitly in CorrelationCholesky bijector.
The bijector base class definition is not able to handle rank-changing bijectors such as CorrelationCholesky. The problem is that the value of event_ndims for fldj is sent to ildj as is but it should be 'event_ndims + 1'.

For this specific case, the resulting exception only gets triggered when value of 'event_ndims' is not known statically -- for instance when invoked by HamiltonianMonteCarlo in Graph mode. The error is still there in Eager mode IIUC but in this case the final answer is correct as no exception is thrown.

I've added a unit test for this case which fails without this change.

PiperOrigin-RevId: 249770795