Open-source the program-counter-based auto-batching system.
This is a system for single-program multiple-data parallelism on batch-oriented hardware.  The main idea is to keep track of which basic block each logical thread wants to execute, and batch together the ones that are in the same place.  The user-facing frontend is a source-to-source Python transformation.  The backend is a lightweight virtual machine written in TensorFlow that simulates user programs in this style.

PiperOrigin-RevId: 250990302