Standardize parameter scales during STS inference.
Currently we see optimization and numerical issues when the `observed_time_series` has large scale (stddev >> 1000): in this case initializing scale params to small values can lead to very large or even NaN initial loss. Conceptually we might fix this by standardizing the input, `(observed_time_series - mean)/stddev`, but doing so would change the interpretation of parameter priors and posterior, and create some awkwardness in ensuring we transform all forecasts, decompositions, etc. back to the original scale.

This CL achieves a roughly equivalent effect, without explicit standardization, by doing inference over scale parameters in a rescaled space. It allows the base variational distribution or sampler to represent a scale parameter in the units of the 'standardized' series, and then multiplies this by the observed stddev to get a value appropriate to the actual series. As with standardization, small initialization and small steps in the inference space now correspond to large initialization / large steps on the original time series.

Note that this only changes the conditioning of the inference problem -- it affects the speed of inference but not the result. The posterior on parameters (and related quantities like the optimal variational parameters, value of the ELBO at the optimum, etc.) is unchanged.

PiperOrigin-RevId: 251346330