Statically detect when a callee doesn't write a variable, and avoid promoting it to the FULL allocation strategy in that case.
On the NUTS program, this leads to a 2x reduction in FULL variables (because build_tree doesn't write any of the same variables as evolve_trajectory), which should yield a close to 2x reduction in memory use by the auto-batched state.

This optimization is also crucial for something like a slice sampler, which has no recursion at all and therefore shouldn't need any stacks.

PiperOrigin-RevId: 251460578