Standardize parameter scales during STS inference. (second try)
Currently we see optimization and numerical issues when the `observed_time_series` has large scale (stddev >> 1000): in this case initializing scale params to small values can lead to very large or even NaN initial loss. Conceptually we might fix this by standardizing the input, `(observed_time_series - mean)/stddev`, but doing so would change the interpretation of parameter priors and posterior, and create some awkwardness in ensuring we transform all forecasts, decompositions, etc. back to the original scale.

This CL achieves a roughly equivalent effect, without explicit standardization, by doing inference over scale parameters in a rescaled space. It allows the base variational distribution or sampler to represent a scale parameter in the units of the 'standardized' series, and then multiplies this by the observed stddev to get a value appropriate to the actual series. As with standardization, small initialization and small steps in the inference space now correspond to large initialization / large steps on the original time series.

Note that this only changes the conditioning of the inference problem -- it affects the speed of inference but not the result. The posterior on parameters (and related quantities like the optimal variational parameters, value of the ELBO at the optimum, etc.) is unchanged.

An earlier version of this change, cl/251346330, applied the rescaling to all scale params automatically in the inference code, instead of explicitly in each component. That was a more confined change, but the current approach has some advantages:
 - All scale parameters see the same scaling, exactly determined by the observed stddev (as opposed to auto-scale based on the prior stddev, which varies between parameters). This gives more exact equivalence with the naive approach of standardizing the observed time series.
 - Fewer unintended consequences. Some 'scale' parameters don't need to be rescaled and might not even have priors with a well-defined stddev (in particular, the previous attempt broke SparseLinearRegression which uses InvGamma(.5, .5) priors). Controlling the scaling explicitly in each component avoids this problem so we only rescale the parameters we really mean to.

A downside is worse separation of model from inference: any new components will need to implement rescaling manually, and forgetting to do this will yield weird behavior from partially-rescaled inference. But this seems like not-a-big-deal because anyone implementing a new component will use existing components as examples.

PiperOrigin-RevId: 251955595