Add a pop-push fusion pass to auto-batching.
A stack pop followed by a stack push (with no intervening read) is equivalent to just updating the top of the stack.  The latter is more efficient for FULL variables, because it just updates the cache for the top, and avoids gathering from and scattering to the backing stack Tensor.  The new pass finds such sequences and replaces them with explicit updates.  This reduces scatter-gather pressure.

To accommodate the optimization, also add a `skip_push_mask` field to `PrimOp`.  This mask identifies those variables the `PrimOp` writes that should be updated in place rather than pushed.  The underlying `Variable`s already implement both pushing and updating.

PiperOrigin-RevId: 252496572