Clean up stack manipulation of the program counter.
The way the code was before was a relic of two things:

- `PushGotoOp` used to be the sequence `PushAddressOp` and `GotoOp`,
  and the code had to be arranged so that `PushAddressOp` in
  particular would write the target address into the correct level of
  the stack.

- `Environment` didn't used to have a method for just updating
  variables.

Given that both of these changed, the code makes more sense this way.
It will also touch the program counter's stack much less, which can be
relevant for performance on while loops that don't have much other
stack interaction going on.

PiperOrigin-RevId: 253029597