Lower self-calls to push-pop more carefully.
This has two effects: it optimizes self-writes (*) to do zero work; and it detects swap-writes (**) and emits a compile-time error instead of silently computing the wrong answer.

(*) A self-write is a function calling itself with an unchanged parameter:
def gcd(x, y):
  ...
  return gcd(x, mod(y, x))

Here `x` should be just passed through, with no stack pop or push, and not even a `where` operation.  This CL achieves this.

(**) A swap-write is a function calling itself permuting its parameters:
def gcd(x, y):
  ...
  return gcd(y, x)

In the naive lowering, this would lead to trying to swap `x` and `y` on the tops of their stacks with just one push and pop operation, which is not possible.  This situation is now flagged as a compile-time error.  Workaround is to add a temporary:
def gcd(x, y):
  ...
  tmp_y = y
  return gcd(tmp_y, x)

PiperOrigin-RevId: 253388508