Normalize scipy usage in tests (and one random use of `mock`).
Early on we established a pattern of allowing tests to pass even without certain deps (by bypassing those tests when the deps weren't present). Meanwhile usage patterns have drifted and we're inconsistent in how we handle this. Since it's quite low friction to install our handful of test deps, and users can just use Travis if they really don't want to bother, we are moving to 1. be consistent and 2. err on the side of demanding deps to be present.

PiperOrigin-RevId: 253915355