Pin to day old tf-nightly. Also, make it easier to do this.
Specifically, create a 'bad dates' regex in the test scripts, and then do some hackery to find the newest tf-nightly[-2.0-preview] version not on that list. Getting pip to tell you what versions are available is apparently not that easy (see https://stackoverflow.com/questions/4888027/python-and-pip-list-all-versions-of-a-package-thats-available). The approach chosen is not something I'm proud of, but will work.

PiperOrigin-RevId: 258428500