Support TF2/Eager-mode fitting of STS models, and deprecate `build_factored_variational_loss`.
This CL updates STS fitting to use modern TF2/Eager-friendly idioms. We rely on recent changes in TFP supporting tapeable distributions (using DeferredTensor and relying on the Normal distribution not to convert variables to tensors in init), to add a new method `build_factored_surrogate_posterior`, which returns a trainable joint distribution over model parameters. This surrogate posterior can be fit in one line using the recently-added `tfp.vi.fit_surrogate_posterior`, removing the need for users to write their own optimization loops.

As a bonus, we inherit the functionality of `tfp.vi.fit_surrogate_posterior`, including support for multiple Monte Carlo samples and alternative variational divergences.

The previous method for VI fitting, `build_factored_variational_loss`, created variables defining a surrogate posterior *and* used them to build a variational loss. This was a poor fit for Eager-mode semantics, where we want to create the variables outside of a GradientTape, but build the loss inside the tape; in particular it led to several issues
https://github.com/tensorflow/probability/issues/452
https://github.com/tensorflow/probability/issues/457
all fixed by this CL.

NOTE: this CL preserves the previous `build_factored_variational_loss` API, so it should not be breaking for most users, but it does break XLA compilation of that method because it simplifies the construction of internal variables to use a direct initialization instead of deferred initialization. The solution is to build the variables outside of the compiled training block using the new-style API. I've included fixes for the known Google3 users; I'll run a global TAP train before submitting.
PiperOrigin-RevId: 260808918