Support new TF2 dtype behavior in Keras layers.
In TensorFlow 2, layers will default to float32 and cast inputs to the layer's dtype. This means many existing float64 models will be silently start using float32 in TF 2 (although a warning is issued).

This change makes code in tensorflow_probability/python/layers support this new behavior. In particular, whenever one layer creates another layer, the outer layer now passes the correct dtype to the inner layer.

It also fixes all tests in tensorflow_probability/python/layers to run in the same dtype as in TF 1. Doing this mostly consisted of passing "dtype=self.dtype" to layer constructors. I also could have set floatx to float64, which would have been simpler, but then I would not test that outer layers pass the correct dtype to inner layers.

In order to ensure the dtypes in the test are the same as in TF 1, I manually checked that the warning TF 2 issues when the new behavior is encountered does not appear. Some assertions checking dtypes are also added.

PiperOrigin-RevId: 261247104